Lower the Python requirement
Looking at the Python 3.3 release notes, there shouldn't be anything we
need that isn't in 3.2.3

Many distributions still ship with it, so lowering this requirement is
an easy win towards easier adoption by more distributions.

This is as a commitment on our part to support installation of pycangjie
on Python 3.2.3 too. If something works on 3.3 but doesn't on 3.2.3, we
will consider it a bug and fix it.