Check for the minimum version of Cython
Cython 0.17.3 is what I've been using all along to develop ibus-cangjie,
so it is known to work. (however, see bug #16 for how we might want to
lower that)

I couldn't find any nice way to check for a minimum version with
Autoconf, but if we get to this point in the configure script, the we
know that we have Python already, so let's use that.