tests: Ensure pycangjie provides expected results
We do not want test that the results from pycangjie are actually
correct, only that they are are identical to the results from
libcangjie.

Correctness must be guaranteed by libcangjie, so that everybody
benefits.

The tests implemented here rely on libcangjie_cli: we run it for every
possible input code, and check that we obtain the same results with the
pycangjie API.

libcangjie_cli is linked directly against libcangjie, and is provided by
libcangjie upstream, so we should be able to trust its results.

However, we scrape its standard output to compare with our results,
which is quite fragile, but it is all we have at the moment.